+++ otherbins/parsestats4DROMPAplus.pl
@@ -33,22 +33,24 @@ while(<File>){
 	if($gcsummit eq ""){
 	    $depth = $clm[14];
 	    $total_gc_base = $clm[17];
-#	    $FRiP = $clm[18] if($#clm >= 18);
+	    $nread_in_peak = $clm[19];
+	    $FRiP = $clm[20];
 	}else{
 	    $depth = $clm[17];
 	    $total_gc_base = $clm[20];
-#	    $FRiP = $clm[21] if($#clm >= 21);
+	    $nread_in_peak = $clm[22];
+	    $FRiP = $clm[23];
 	}
     }
 }
 close (File);
 
-print STDERR "Sample\tMapped reads\t + strand\t - strand\tRedundancy threshold\tNonredundant\tRedundant\tComplexity for10M\tRead depth\tGenome coverage\tTested_reads";
-print STDERR "\tGC summit" if($gcsummit ne "");
-print STDERR "\tFRiP" if($FRiP ne "");
-print STDERR "\n";
+print "Sample\tMapped reads\t + strand\t - strand\tRedundancy threshold\tNonredundant\tRedundant\tComplexity for10M\tRead depth\tGenome coverage\tTested_reads";
+print "\tGC summit" if($gcsummit ne "");
+print "\treads in peaks\tFRiP";
+print "\n";
 
 print "$file\t$total_reads\t$plus\t$minus\t$pcrbiasthre\t$total_remained\t$total_filtered\t$tested_complexity\t$depth\t$total_gc_base\t$tested_reads";
 print "\t$gcsummit" if($gcsummit ne "");
-print "\t$FRiP" if($FRiP ne "");
+print "\t$nread_in_peak\t$FRiP";
 print "\n";