+++ ChangeLog.md
@@ -2,6 +2,7 @@
 
 ## 1.15.2 (2022-08-05)
 - Bug fix: remove "peak coverage" header from stats file when calculating FRiP score.
+- Modify `otherbins/parsestats4DROMPAplus.pl` to output FRiP score always.
 
 ## 1.15.1 (2022-07-27)
 - Add error messages when the format of supplied genometable/mptable files is wrong.